refactor: 閲覧用フロアマップで部屋情報と滞在者情報を取得して型変換する処理をroomState.tsに切り出した。(useEffectは無限レンダリングを防ぐため必要)
